move everything logging related to new logging folder
