+++ src/autopas/utils/AutoPasConfigurationCommunicator.cpp
@@ -9,7 +9,7 @@
 #include "autopas/containers/CompatibleLoadEstimators.h"
 #include "autopas/containers/CompatibleTraversals.h"
 #include "autopas/utils/ConfigurationAndRankIteratorHandler.h"
-#include "autopas/utils/Logger.h"
+#include "utils/logging/Logger.h"
 
 namespace autopas::utils::AutoPasConfigurationCommunicator {
